@@ -1,9 +1,9 @@
 {
   "private": true,
-  "name": "react-nesting-example-legacy",
+  "name": "react-nesting-example-modern",
   "dependencies": {
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-redux": "7.2.9",
     "react-router-dom": "5.3.4"
   }
