@@ -6,17 +6,17 @@
  */
 
 import React from 'react';
-import {StrictMode} from 'react';
-import ReactDOM from 'react-dom';
-import {Provider} from 'react-redux';
+import { StrictMode } from 'react';
+import { createRoot } from 'react-dom/client';
+import { Provider } from 'react-redux';
 import App from './App';
-import {store} from '../store';
+import { store } from '../store';
 
-ReactDOM.render(
+const root = createRoot(document.getElementById('root'));
+root.render(
   <StrictMode>
     <Provider store={store}>
       <App />
     </Provider>
-  </StrictMode>,
-  document.getElementById('root')
+  </StrictMode>
 );
