@@ -5,4 +5,4 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import './modern/index';
+import './host/index';
