@@ -9,37 +9,30 @@ import React from 'react';
 import { Component } from 'react';
 import { Link } from 'react-router-dom';
 import { connect } from 'react-redux';
-import { store } from '../store';
-
 import ThemeContext from './shared/ThemeContext';
-import Clock from './shared/Clock';
 
-store.subscribe(() => {
-  console.log('Counter:', store.getState());
-});
-
-class GreetingSection extends Component {
+class TextingSection extends Component {
   render() {
     return (
       <ThemeContext.Consumer>
         {(theme) => (
           <div style={{ border: '1px dashed black', padding: 20 }}>
-            <h3>src/legacy/Greeting.js</h3>
+            <h3>src/legacy/Texting.js</h3>
             <h4 style={{ color: theme }}>
               This component is rendered by the nested React ({React.version}).
             </h4>
-            <Clock />
-            <p>
-              Counter: {this.props.counter}{' '}
-              <button
-                onClick={() => this.props.dispatch({ type: 'increment' })}
-              >
-                +
-              </button>
-            </p>
-            <b>
-              <Link to='/'>Go to Home</Link>
-            </b>
+            <span>Text from global state: </span>
+            <input
+              value={this.props.text}
+              onChange={(e) =>
+                this.props.dispatch({ type: 'text', text: e.target.value })
+              }
+            ></input>
+            <br />
+
+            <Link to='/'>Go to Home</Link>
+            <br />
+            <Link to='/greeting'>Go to GreetingPage</Link>
           </div>
         )}
       </ThemeContext.Consumer>
@@ -49,9 +42,8 @@ class GreetingSection extends Component {
 
 function mapStateToProps(state) {
   return {
-    counter: state.counter,
     text: state.text,
   };
 }
 
-export default connect(mapStateToProps)(GreetingSection);
+export default connect(mapStateToProps)(TextingSection);
